requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile pyproject.toml
#
cloudpickle==3.1.0
    # via gymnasium
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
farama-notifications==0.0.4
    # via gymnasium
fonttools==4.55.3
    # via matplotlib
gymnasium==1.0.0
    # via graph-matrix-jsp-env (pyproject.toml)
jsp-vis==1.1.0
    # via graph-matrix-jsp-env (pyproject.toml)
kaleido==0.2.1
    # via jsp-vis
kiwisolver==1.4.8
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
matplotlib==3.8.4
    # via
    #   graph-matrix-jsp-env (pyproject.toml)
    #   jsp-vis
mdurl==0.1.2
    # via markdown-it-py
numpy==2.2.1
    # via
    #   contourpy
    #   graph-matrix-jsp-env (pyproject.toml)
    #   gymnasium
    #   jsp-vis
    #   matplotlib
    #   opencv-python
    #   pandas
opencv-python==4.10.0.84
    # via jsp-vis
packaging==24.2
    # via
    #   matplotlib
    #   plotly
pandas==2.2.3
    # via jsp-vis
pillow==11.1.0
    # via matplotlib
plotly==5.24.1
    # via jsp-vis
pygments==2.19.1
    # via rich
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.2
    # via pandas
rich==13.9.4
    # via graph-matrix-jsp-env (pyproject.toml)
six==1.17.0
    # via python-dateutil
tenacity==9.0.0
    # via plotly
typing-extensions==4.12.2
    # via gymnasium
tzdata==2024.2
    # via pandas
